Update map geometry when tracking changes
This updates the map center or region when either the current location
or the filter distance respectively change. But, the map is only
updated when tracking the current location.

Note, this commit actually removes some behavior. In my defense, there
was a comment that said:

"// Just zoom to the new search radius (or maybe don't even do that?)"

I agree with the doubt, so I removed that behavior.